requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
python-dotenv==1.0.1
litellm==1.52.5
langgraph==0.2.41
chromadb==0.5.11
qdrant-client==1.12.1
langfuse==2.48.2
rg-lib==0.0.2
requests==2.32.3
typer==0.12.5
rich==13.9.3
orjson==3.10.7


